Improved token location recording in the syntax tree to now include the start line and column of nodes in the parse tree

+++ src/parser/errors.rs
@@ -5,7 +5,7 @@ use super::token::Token;
 
 #[derive(Debug, Clone)]
 pub enum LexingError {
-    UnrecognizedToken(String, u64, u64),
+    UnrecognizedToken(String, usize, usize),
     InvalidEscapeCharacter(char),
     NonTerminatedString
 }
